+++ udata/__init__.py
@@ -4,5 +4,5 @@
 uData
 '''
 
-__version__ = '2.1.0'
+__version__ = '2.1.1.dev'
 __description__ = 'Open data portal'