+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Current (in progress)
+
+- Nothing yet
+
 ## 2.1.0 (2020-05-12)
 
 ### Breaking changes